fix(LottoTest): 패키지 변경으로 인한 테스트 코드 오류

+++ src/main/java/lotto/controller/LotteryMachine.java
@@ -32,7 +32,7 @@ public class LotteryMachine {
         final EnumMap<Grade, Integer> gradeWithCount = judge.allJudgedGrade();
         OutputView.printResultScore(gradeWithCount);
 
-        final Long profit = profitCalculator.calculateProfit(gradeWithCount, purchasePrice);
+        final Double profit = profitCalculator.calculateProfit(gradeWithCount, purchasePrice);
         OutputView.printProfitPercentage(profit);
     }
