+++ src/main/java/lotto/model/Calculator.java
@@ -1,7 +1,7 @@
 package lotto.model;
 
 public abstract class Calculator {
-    long divide(int divided, int dividend) {
-        return divided / dividend;
+    Double divide(Integer divided, Integer dividend) {
+        return (double) ((double) divided / dividend);
     }
 }
